third_party/WebKit/LayoutTests/http/tests/htmlimports/import-blocking-nested-child.html

<!DOCTYPE html>
<html>
<head>
<script src="/js-test-resources/js-test-pre.js"></script>
<link id="leaf" rel="import" href="resources/import-blocking-nested-child-1.html">
</head>
<body>
<script>
shouldBeTrue("blockingScriptHasRun");
</script>
</body>
</html>
